test: split remote backend lifecycle coverage
